Async method for common variable method
