+++ source/Octopus.Client.Tests/PublicSurfaceAreaFixture.ThePublicSurfaceAreaShouldNotRegress..NETFramework.approved.txt
@@ -7319,6 +7319,37 @@ Octopus.Client.Model.Migrations
     String TaskId { get; set; }
   }
 }
+Octopus.Client.Model.TenantVariables
+{
+  class CommonVariableScope
+  {
+    .ctor(String[])
+    String[] EnvironmentIds { get; set; }
+  }
+  class ModifyCommonVariablesByTenantIdCommand
+  {
+    .ctor(String, String, Octopus.Client.Model.TenantVariables.TenantCommonVariable[])
+    String SpaceId { get; set; }
+    String TenantId { get; set; }
+    Octopus.Client.Model.TenantVariables.TenantCommonVariable[] Variables { get; set; }
+  }
+  class ModifyCommonVariablesByTenantIdResponse
+  {
+    .ctor(String, Octopus.Client.Model.TenantVariables.TenantCommonVariable[])
+    Octopus.Client.Model.TenantVariables.TenantCommonVariable[] CommonVariables { get; set; }
+    String TenantId { get; set; }
+  }
+  class TenantCommonVariable
+  {
+    .ctor(String, String, Octopus.Client.Model.PropertyValueResource, Octopus.Client.Model.TenantVariables.CommonVariableScope)
+    .ctor(String, String, String, Octopus.Client.Model.PropertyValueResource, Octopus.Client.Model.TenantVariables.CommonVariableScope)
+    String Id { get; set; }
+    String LibraryVariableSetId { get; set; }
+    Octopus.Client.Model.TenantVariables.CommonVariableScope Scope { get; set; }
+    String TemplateId { get; set; }
+    Octopus.Client.Model.PropertyValueResource Value { get; set; }
+  }
+}
 Octopus.Client.Model.Triggers
 {
   class ArcFeedFilterResource
@@ -7821,11 +7852,6 @@ Octopus.Client.Repositories
     Octopus.Client.Model.ResourceCollection<OctopusPackageVersionBuildInformationMappedResource> ListBuilds(String, Int32, Int32)
     Octopus.Client.Model.BuildInformation.OctopusPackageVersionBuildInformationMappedResource Push(String, String, Octopus.Client.Model.BuildInformation.OctopusBuildInformation)
     Octopus.Client.Model.BuildInformation.OctopusPackageVersionBuildInformationMappedResource Push(String, String, Octopus.Client.Model.BuildInformation.OctopusBuildInformation, Octopus.Client.Model.OverwriteMode)
-  }
-  class CommonVariableScope
-  {
-    .ctor(String[])
-    String[] EnvironmentIds { get; set; }
   }
   class DeploymentFreezeRepository
     Octopus.Client.Repositories.IDeploymentFreezeRepository
@@ -8465,6 +8491,7 @@ Octopus.Client.Repositories
     Octopus.Client.Repositories.IGetAll<TenantVariableResource>
   {
     List<TenantVariableResource> GetAll(Octopus.Client.Model.ProjectResource)
+    Octopus.Client.Model.TenantVariables.ModifyCommonVariablesByTenantIdResponse Modify(Octopus.Client.Model.TenantVariables.ModifyCommonVariablesByTenantIdCommand)
   }
   interface IUpgradeConfigurationRepository
     Octopus.Client.Repositories.IGet<UpgradeConfigurationResource>
@@ -8561,19 +8588,6 @@ Octopus.Client.Repositories
     Octopus.Client.Model.MachineConnectionStatus GetConnectionStatus(Octopus.Client.Model.WorkerResource)
     Octopus.Client.Model.ResourceCollection<WorkerResource> List(Int32, Nullable<Int32>, String, String, String, Nullable<Boolean>, String, String, String)
   }
-  class ModifyCommonVariablesByTenantIdCommand
-  {
-    .ctor(String, String, Octopus.Client.Repositories.TenantCommonVariable[])
-    String SpaceId { get; set; }
-    String TenantId { get; set; }
-    Octopus.Client.Repositories.TenantCommonVariable[] Variables { get; set; }
-  }
-  class ModifyCommonVariablesByTenantIdResponse
-  {
-    .ctor(String, Octopus.Client.Repositories.TenantCommonVariable[])
-    Octopus.Client.Repositories.TenantCommonVariable[] CommonVariables { get; set; }
-    String TenantId { get; set; }
-  }
   class PerformanceConfigurationRepository
     Octopus.Client.Repositories.IPerformanceConfigurationRepository
   {
@@ -8589,16 +8603,6 @@ Octopus.Client.Repositories
     Octopus.Client.Model.TelemetryConfigurationResource EnableTelemetry()
     Octopus.Client.Model.TelemetryConfigurationResource GetTelemetryConfiguration()
     Octopus.Client.Model.TelemetryConfigurationResource ModifyTelemetryConfiguration(Octopus.Client.Model.TelemetryConfigurationResource)
-  }
-  class TenantCommonVariable
-  {
-    .ctor(String, String, Octopus.Client.Model.PropertyValueResource, Octopus.Client.Repositories.CommonVariableScope)
-    .ctor(String, String, String, Octopus.Client.Model.PropertyValueResource, Octopus.Client.Repositories.CommonVariableScope)
-    String Id { get; set; }
-    String LibraryVariableSetId { get; set; }
-    Octopus.Client.Repositories.CommonVariableScope Scope { get; set; }
-    String TemplateId { get; set; }
-    Octopus.Client.Model.PropertyValueResource Value { get; set; }
   }
 }
 Octopus.Client.Repositories.Async
@@ -9347,6 +9351,7 @@ Octopus.Client.Repositories.Async
     Octopus.Client.Repositories.Async.IGetAll<TenantVariableResource>
   {
     Task<List<TenantVariableResource>> GetAll(Octopus.Client.Model.ProjectResource)
+    Task<ModifyCommonVariablesByTenantIdResponse> Modify(Octopus.Client.Model.TenantVariables.ModifyCommonVariablesByTenantIdCommand, CancellationToken)
   }
   interface IUpgradeConfigurationRepository
     Octopus.Client.Repositories.Async.IGet<UpgradeConfigurationResource>
